[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.03 ms, Memory: 72.7 MB -BaekjoonHub

+++ 프로그래머스/0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -1,5 +1,5 @@
 class Solution {
-    public double solution(int balls, int share) {
+    public int solution(int balls, int share) {
         
         double answer = 1;
 
@@ -8,6 +8,6 @@ class Solution {
             answer /= (i-share);
         }
 
-        return answer;
+        return (int)answer;
     }
 }
